+++ tests/test_stage4.py
@@ -1,0 +1,8 @@
+"""Test stage4 information."""
+
+from pyiem import stage4
+
+
+def test_api():
+    """Test API exists."""
+    assert stage4.DY